汎用的に使うメソッドをutilに移動
