+++ kit_timetable_rebuilder/__init__.py
@@ -1,2 +1,3 @@
 from .lesson import *
 from .timetable_parser import *
+from .util import *
